+++ stream-chat-android-ui-components-sample/src/main/res/layout/fragment_channels.xml
@@ -16,7 +16,7 @@
         app:layout_constraintTop_toTopOf="parent"
         />
 
-    <io.getstream.chat.android.ui.channel.list.ChannelsView
+    <io.getstream.chat.android.ui.channel.list.ChannelListView
         android:id="@+id/channelsView"
         android:layout_width="0dp"
         android:layout_height="0dp"
