+++ stream-chat-android-ui-components-sample/src/main/res/layout/fragment_chat_info_shared_groups.xml
@@ -26,7 +26,7 @@
 
     </androidx.appcompat.widget.Toolbar>
 
-    <io.getstream.chat.android.ui.channel.list.ChannelsView
+    <io.getstream.chat.android.ui.channel.list.ChannelListView
         android:id="@+id/channelsView"
         android:layout_width="0dp"
         android:layout_height="0dp"
